Add recieving survey questions

+++ src/quizer/application/interactors/survey/delete_survey.py
@@ -1,6 +1,7 @@
 from uuid import UUID
 from quizer.application.interfaces.repositories.survey import SurveyRepository
 from quizer.application.interfaces.common.id_provider import IdProvider
+from quizer.application.exceptions import TargetNotFoundError
 
 
 class DeleteSurveyInteractor:
@@ -11,5 +12,7 @@ class DeleteSurveyInteractor:
     async def __call__(self, survey_id: UUID) -> None:
         user_id = self._id_provider.get_current_user_id()
         survey = await self._survey_repo.get_by_id(survey_id)
+        if survey is None:
+            raise TargetNotFoundError("Survey was not found")
         survey.can_manage(user_id)
         await self._survey_repo.delete(survey_id)
